+++ twidere/src/main/kotlin/org/mariotaku/twidere/fragment/ParcelableStatusesFragment.kt
@@ -147,7 +147,7 @@ abstract class ParcelableStatusesFragment : AbsStatusesFragment() {
     override fun onLoadMoreContents(position: Long) {
         // Only supports load from end, skip START flag
         if (position and ILoadMoreSupportAdapter.START !== 0L || refreshing) return
-        super.onLoadMoreContents(position.toLong())
+        super.onLoadMoreContents(position)
         if (position == 0L) return
         val adapter = adapter
         // Load the last item
